Allows Donors to vary the amount of bonus tickets used
Typically, the amount from your pool that you take out on a given round to get antag is forced at the config value. This alters that, now allowing Donators to chose to bid lower amounts of tickets, to perhaps conserve them and prioritize getting antag when they really want it.